+++ api/errors.py
@@ -1,0 +1,2 @@
+class RequestError(Exception):
+    pass
